Converted build from CRA to Vite
