--- public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "App",
-  "name": "React App",
-  "icons": [
-    {
-      "src": "dice.png",
-      "type": "image/png",
-      "sizes": "512x512"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
